simulation-module/requirements.txt
pyspark==3.5.0 
kafka-python 
pandas 
numpy 
psycopg2-binary
matplotlib
setuptools 
wheel